Major refactoring.
This more closely resembles the JSON implementation: parse into an
in-memory representation and then decode on that. This is probably
a major perf hit, but doing the reading directly in the decoder is
impossible or very hard.

Basically, in order to satisfy `Decoder`, you really need to keep a
stack of values around that can be manipulated via the various
continuations. This stack is the in-memory representation, and it
includes heap allocation and boxing every primitive value.

Another approach is to maintain a stack, but in CBOR encoded form.
This would probably be much faster, but is definitely a "v2" feature.
(And by then, we might have an entirely different serialization
infrastructure!)