Fix CBOR tag encoding.
The contortions required to support CBOR tag encoding ended up biting
me here. It turned out that my approach what mistakenly ignoring
encoding map field names when they were part of a map being encoded
as part of a CBOR tag. I've fixed this by explicitly checking for the
struct member names of a `CborTagEncode`. This required making the field
names private and adding a `new` constructor in its place.

Therefore, change code that looks like this:

    CborTagEncode { tag: 100, data: val }.encode(e)

to this:

    CborTagEncode::new(100, val).encode(e)

This should only affect users that are encoding custom CBOR tags.

[breaking-change]