Fixes #6.
When reading the next data item, the only time EOF is expected is when
reading the first byte. If EOF occurs after that while reading an item,
then it is unexpected and it should be reported as an error.

Previously, all EOFs---regardless of whether they were expected or
not---were be swallowed by the iterator and translated to `None`
values.